[BE] feat : create spot

+++ BE/tripeer/src/main/java/j10d207/tripeer/place/db/entity/SpotDetailEntity.java
@@ -17,7 +17,7 @@ public class SpotDetailEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     // PK
-    private long spotDetailId;
+    private int spotDetailId;
 
     @OneToOne
     @MapsId
